requirements.txt
requests>=2.3.0
urllib3>=1.24.1
PyJWT>=1.7.1